fix(web): keep resolveBillingLink out of the inactive page's server module
SvelteKit only permits its own names as exports from +page.server.ts, so
exporting the helper for its unit test made `vite build` refuse the route.
That broke the Docker web image build on main and on every open PR since
#1184 landed. The helper now lives in a sibling module, matching how
recovery-error.ts and activation-error.ts are arranged, and the test
imports it from there.
